+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 7.55
 -----
 
+*   Updates:
+    *   Add theme support to Up Next screen
+        ([#1605](https://github.com/Automattic/pocket-casts-android/pull/1605))
 
 7.54
 -----